requirements.txt
streamlit==1.27.0
anthropic==0.20.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
pdf2image==1.16.3
opencv-python-headless==4.8.0.76